+++ src/index.ts
@@ -5,3 +5,4 @@ export * from "./utils/getDataFieldsFromName";
 export * from "./utils/getFieldFromName";
 export * from "./utils/getSeriesFromName";
 export * from "./utils/evaluateString";
+export * from "./utils/getMetricValueWithFilters";